[MLOP-225] KeyFeature need dtype to be a required arg (#126)

+++ tests/unit/butterfree/core/load/conftest.py
@@ -13,7 +13,9 @@ from butterfree.core.transform.transformations import AggregatedTransform
 
 @fixture
 def feature_set():
-    key_features = [KeyFeature(name="id", description="Description")]
+    key_features = [
+        KeyFeature(name="id", description="Description", dtype=DataType.INTEGER)
+    ]
     ts_feature = TimestampFeature(from_column=TIMESTAMP_COLUMN)
     features = [
         Feature(name="feature", description="Description", dtype=DataType.BIGINT,)
